requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.3.2
aiohttp>=3.9.0
asyncio>=3.4.3 